Fix typo in package name: contenpacks -> contentpacks
Fixes #6882
