+++ graylog2-server/src/main/java/org/graylog2/rest/models/system/contentpacks/responses/CatalogResolveRequest.java
@@ -14,26 +14,25 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.rest.models.system.contenpacks.responses;
+package org.graylog2.rest.models.system.contentpacks.responses;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
-import org.graylog.autovalue.WithBeanGetter;
+import com.google.common.collect.ImmutableSet;
+import org.graylog2.contentpacks.model.entities.EntityDescriptor;
 
-import java.util.Map;
+import java.util.Set;
 
 @JsonAutoDetect
-
 @AutoValue
-@WithBeanGetter
-public abstract class ContentPackMetadata {
-    @JsonProperty
-    public abstract int installationCount();
+public abstract class CatalogResolveRequest {
+    @JsonProperty("entities")
+    public abstract ImmutableSet<EntityDescriptor> entities();
 
     @JsonCreator
-    public static ContentPackMetadata create(@JsonProperty("installation_count") int installationCount) {
-        return new AutoValue_ContentPackMetadata(installationCount);
+    public static CatalogResolveRequest create(@JsonProperty("entities") Set<EntityDescriptor> entities) {
+        return new AutoValue_CatalogResolveRequest(ImmutableSet.copyOf(entities));
     }
 }
