+++ graylog2-server/src/main/java/org/graylog2/rest/models/system/contentpacks/responses/CatalogResolveResponse.java
@@ -14,24 +14,25 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.rest.models.system.contenpacks.responses;
+package org.graylog2.rest.models.system.contentpacks.responses;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
-import org.graylog2.contentpacks.model.entities.EntityExcerpt;
+import com.google.common.collect.ImmutableSet;
+import org.graylog2.contentpacks.model.entities.Entity;
 
-import java.util.Set;
+import java.util.Collection;
 
 @JsonAutoDetect
 @AutoValue
-public abstract class CatalogIndexResponse {
+public abstract class CatalogResolveResponse {
     @JsonProperty("entities")
-    public abstract Set<EntityExcerpt> entities();
+    public abstract ImmutableSet<Entity> entities();
 
     @JsonCreator
-    public static CatalogIndexResponse create(@JsonProperty("entities") Set<EntityExcerpt> entities) {
-        return new AutoValue_CatalogIndexResponse(entities);
+    public static CatalogResolveResponse create(@JsonProperty("entities") Collection<Entity> entities) {
+        return new AutoValue_CatalogResolveResponse(ImmutableSet.copyOf(entities));
     }
 }
