+++ graylog2-server/src/main/java/org/graylog2/rest/models/system/contentpacks/responses/ContentPackInstallationRequest.java
@@ -14,26 +14,35 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.rest.models.system.contenpacks.responses;
+package org.graylog2.rest.models.system.contentpacks.responses;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
-import com.google.common.collect.ImmutableSet;
-import org.graylog2.contentpacks.model.constraints.Constraint;
-import org.graylog2.contentpacks.model.entities.Entity;
+import org.graylog.autovalue.WithBeanGetter;
+import org.graylog2.contentpacks.model.entities.references.ValueReference;
 
-import java.util.Collection;
+import javax.annotation.Nullable;
+import java.util.Collections;
+import java.util.Map;
 
 @JsonAutoDetect
 @AutoValue
-public abstract class CatalogResolveResponse {
-    @JsonProperty("entities")
-    public abstract ImmutableSet<Entity> entities();
+@WithBeanGetter
+public abstract class ContentPackInstallationRequest {
+    @JsonProperty("parameters")
+    public abstract Map<String, ValueReference> parameters();
+
+    @JsonProperty("comment")
+    @Nullable
+    public abstract String comment();
 
     @JsonCreator
-    public static CatalogResolveResponse create(@JsonProperty("entities") Collection<Entity> entities) {
-        return new AutoValue_CatalogResolveResponse(ImmutableSet.copyOf(entities));
+    public static ContentPackInstallationRequest create(
+            @JsonProperty("parameters") @Nullable Map<String, ValueReference> parameters,
+            @JsonProperty("comment") @Nullable String comment) {
+        final Map<String, ValueReference> parameterMap = parameters == null ? Collections.emptyMap() : parameters;
+        return new AutoValue_ContentPackInstallationRequest(parameterMap, comment);
     }
 }
