+++ graylog2-server/src/main/java/org/graylog2/rest/models/system/contentpacks/responses/ContentPackRevisions.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog2.rest.models.system.contenpacks.responses;
+package org.graylog2.rest.models.system.contentpacks.responses;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -22,7 +22,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
 import org.graylog.autovalue.WithBeanGetter;
 import org.graylog2.contentpacks.model.ContentPack;
-import org.graylog2.contentpacks.model.ModelId;
+import org.graylog2.contentpacks.model.constraints.ConstraintCheckResult;
 
 import java.util.Map;
 import java.util.Set;
@@ -31,20 +31,16 @@ import java.util.Set;
 
 @AutoValue
 @WithBeanGetter
-public abstract class ContentPackList {
-    @JsonProperty
-    public abstract long total();
+public abstract class ContentPackRevisions {
+    @JsonProperty("content_pack_revisions")
+    public abstract Map<Integer, ContentPack> contentPackRevisions();
 
-    @JsonProperty
-    public abstract Set<ContentPack> contentPacks();
-
-    @JsonProperty
-    public abstract Map<ModelId, Map<Integer, ContentPackMetadata>> contentPacksMetadata();
+    @JsonProperty("constraints_result")
+    public abstract Map<Integer, Set<ConstraintCheckResult>> constraints();
 
     @JsonCreator
-    public static ContentPackList create(@JsonProperty("total") long total,
-                                         @JsonProperty("content_packs")Set<ContentPack> contentPacks,
-                                         @JsonProperty("content_pack_metadata") Map<ModelId, Map<Integer, ContentPackMetadata>> contentPacksMetadata) {
-        return new AutoValue_ContentPackList(total, contentPacks, contentPacksMetadata);
+    public static ContentPackRevisions create(@JsonProperty("content_pack_revisions") Map<Integer, ContentPack> contentPackRevisions,
+                                              @JsonProperty("constraints_result")Map<Integer, Set<ConstraintCheckResult>> constraints) {
+        return new AutoValue_ContentPackRevisions(contentPackRevisions, constraints);
     }
 }
