+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/contentpacks/ContentPackResource.java
@@ -40,12 +40,12 @@ import org.graylog2.contentpacks.model.ModelId;
 import org.graylog2.contentpacks.model.Revisioned;
 import org.graylog2.contentpacks.model.constraints.ConstraintCheckResult;
 import org.graylog2.plugin.database.users.User;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackInstallationRequest;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackInstallationsResponse;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackList;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackMetadata;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackResponse;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackRevisions;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackInstallationRequest;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackInstallationsResponse;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackList;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackMetadata;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackResponse;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackRevisions;
 import org.graylog2.shared.rest.resources.RestResource;
 import org.graylog2.shared.security.RestPermissions;
 import org.slf4j.Logger;
@@ -66,7 +66,6 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import java.net.URI;
-import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
 import java.util.function.Function;
