+++ graylog2-server/src/test/java/org/graylog2/rest/resources/system/contentpacks/CatalogResourceTest.java
@@ -31,9 +31,9 @@ import org.graylog2.contentpacks.model.ModelType;
 import org.graylog2.contentpacks.model.entities.EntityDescriptor;
 import org.graylog2.contentpacks.model.entities.EntityExcerpt;
 import org.graylog2.contentpacks.model.entities.EntityV1;
-import org.graylog2.rest.models.system.contenpacks.responses.CatalogIndexResponse;
-import org.graylog2.rest.models.system.contenpacks.responses.CatalogResolveRequest;
-import org.graylog2.rest.models.system.contenpacks.responses.CatalogResolveResponse;
+import org.graylog2.rest.models.system.contentpacks.responses.CatalogIndexResponse;
+import org.graylog2.rest.models.system.contentpacks.responses.CatalogResolveRequest;
+import org.graylog2.rest.models.system.contentpacks.responses.CatalogResolveResponse;
 import org.graylog2.shared.bindings.GuiceInjectorHolder;
 import org.junit.Before;
 import org.junit.Rule;
