+++ graylog2-server/src/test/java/org/graylog2/rest/resources/system/contentpacks/ContentPackResourceTest.java
@@ -25,10 +25,10 @@ import org.graylog2.contentpacks.model.ContentPackInstallation;
 import org.graylog2.contentpacks.model.ModelId;
 import org.graylog2.contentpacks.model.constraints.ConstraintCheckResult;
 import org.graylog2.jackson.AutoValueSubtypeResolver;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackList;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackMetadata;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackResponse;
-import org.graylog2.rest.models.system.contenpacks.responses.ContentPackRevisions;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackList;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackMetadata;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackResponse;
+import org.graylog2.rest.models.system.contentpacks.responses.ContentPackRevisions;
 import org.graylog2.shared.bindings.GuiceInjectorHolder;
 import org.graylog2.shared.bindings.providers.ObjectMapperProvider;
 import org.junit.Before;
